Add organizations and users tables for multi-tenancy
- Add organizations table with subdomain routing, Stripe integration
- Add users table with authentication, roles, email verification
- Migrate all SQL to PostgreSQL-compatible syntax
- Add sqlc queries for organizations and users CRUD
- Add docker-compose.dev.yml for local PostgreSQL
- Add integration test utilities and tenant isolation tests
- Tag integration tests with build constraint (require PostgreSQL)

+++ internal/handler/keyboard/clients.go
@@ -47,8 +47,8 @@ func (h *Handler) ListClients(w http.ResponseWriter, r *http.Request) {
 	// Get paginated clients
 	clients, err := h.queries.ListClientsPaginated(ctx, repository.ListClientsPaginatedParams{
 		Search: search,
-		Offset: offset,
-		Limit:  int64(clientsPageSize),
+		Offset: int32(offset),
+		Limit:  int32(clientsPageSize),
 	})
 	if err != nil {
 		logger.Error("failed to list clients", "error", err)
